include changes to support OpenBSD and prepare NetBSD support
Some systems do not split -lrt and -ldl from the main libc.

Musl libc is one of them, but offers an empty -ldl and -lrt
precisely to avoid these kind of compilation errors.

OpenBSD tried to add a librt.a but it broke libtool-based packages
on multiple OSes[1] so it got reverted, nad -lrt needs to be removed.

Some other adjustments for FreeBSD also work for other BSD, like
packages in /usr/local/ instead of /usr/, TLC libraries...

[1]: https://marc.info/?l=openbsd-tech&m=139559970907629&w=2